Proper fix: invalidate the transformed rect
The previous quick and dirty fix was causing an issue where empty table cells
were not being highlighted when selected.  This made it easy for users to
inadvertently delete cells from tables, not knowing that they were selected.

I looked deeper into this issue and realized that the selectionRectForRepaint
method in RenderBlock does not property map the selection rect to the repaint
container's coordinate space.  In addition, it does not honor the
clipToVisibleContent parameter.

I modified the RenderBlock::selectionRectForRepaint implementation to use
similar logic used in implementations of this method in other classes.

I also realized that RenderBlockSelectionInfo is no longer needed, since
RenderSelectionInfo already does the right thing using selectionRectForRepaint.