Fixed a bug where spell error is marked on nodes with spellcheck set to false.
DRQS 37544856
http://code.google.com/p/chromium/issues/detail?id=167613
According to HTML5 standard, the immediate parent of the first character
of the word, sentence, or other piece of text is the applicable element
that should have its spellcheck attribute checked to determine if
spellcheck should be performed. This patch added this check so that when
spellcheck="false" is set on a node then its content would not be marked
as misspelled.