Push inline elements down when inserting list
DRQS 37507315
When lisfifying a paragraph, if the start of the paragraph is contained within
inline elements, the insertionPos for the list element does not really push the
inline elements down as far as possible (even though the comment says that it
should do so).

This resulted in the issue where when we are changing list type for list
elements that have style spans, the relistified paragraphs do not merge
correctly with each other because the previous sibling is a span element
instead of a list item element.