Fix RenderBlock::additionalMarginStart
The previous implementation looked for the list item, and returned the
marginStart of that list item.  However, marginStart would be invalid if the
box margins have not been calculated.

Now, we simply return the additionalMarginStart of the previous box.  This
essentially is a recursive algorithm, but note that this recursion is
terminated when it encounters a list item, because RenderListItem overrides
additionalMarginStart to simply return the width of the list marker.