Paint selection gaps to mark line endings
DRQS 38038836
If a line does not have a right gap, but the selection extends past the end of
the line, then we should paint a small gap to indicate that the line ending has
been selected.

This is implemented in a new function called 'lineEndingSelectionGap', which is
similar to 'logicalRightSelectionGap', except that it limits the width to five
pixels.

Also, the lastLogicalRight for a block now extends to the last selected line's
logical right, including the line ending gap.