Make line ending highlight work for RTL blocks
The previous logic for line ending highlights only worked for LTR blocks.  This
change makes it work for RTL blocks as well.

The getLineEndingGapLogicalRight method has been replaced with the
getLineEndingGapLogicalLeftAndRight method, which returns *both* the left and
right positions of the line ending gap.

Also, the calculation of lastLogicalLeft and lastLogicalRight for blocks with
inline children has been fixed to handle RTL blocks as well.