Fix selection gaps on before & after side of cells
In the previous code, the before-side of table cells would be filled across the
entire width of the cell, and only if the selection started before the table
cell.  Also, the after-side of table cells would only be filled if the
selection ended after the table cell.

Now, the after-side of the table cells would be filled if anything in selected
at the end of the cell.  This is done by making inlineSelectionGaps and
blockSelectionGaps set an "isAfterSideSelected" flag if the last item in the
block had anything selected.

Also, the before-side of table-cells will now be filled if the firstRootBox is
selected.  The issue where it filled the entire width has been fixed by setting
lastLogicalLeft and lastLogicalRight to the left & right selection coords on
the first line.  This is done by refactoring the old logic of determining
lastLogicalLeft and lastLogicalRight of the last selected line.  This logic is
now in a new method called "getLineSelectionLogicalLeftAndRight".