Add BlockCommand class
This class is intended to be inherited to implement the actual block operation,
for example indenting and outdenting.

It uses the greedy block extend algorithm to extend the selection to contain
the entire block, then lets the derived class format the block extent, by
overriding formatBlockExtent.

If the derived class does not override formatBlockExtent, then BlockCommand
default behavior is to break up the block extent to sibling ranges (in case the
selection ranges through different levels of nesting).  The derived class can
then handle the sibling ranges by overriding formatBlockSiblings.

This class also handles the case where the selection ranges between different
table cells.  Since table cells are not splittable, this class ensures that the
block extent does not cross the table-cell boundary.