Break out of empty list item in table cell
DRQS 38514960
In upstream webkit, we only break out of empty list items on backspace if the
list item is the first editable position in the contenteditable div.

With this change, we also break out of empty list items if the list item is the
first editable position in a table cell.