Adjust cursor position for every autocorrection
DRQS 38913437
This change moves the logic that restore cursor position after applying
auto-correction into the loop so it is executed for every SpellChecking
result.  This way the cursor is placed at the right position for every
undo.  Otherwise the cursor position is only guaranteed to be correct
after all the auto-corrections are undone.